requirements.txt
pandas>=1.5.0
numpy>=1.23.0
pyarrow>=10.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
scipy>=1.9.0
scikit-learn>=1.1.0
# optional: fuzzy_cmeans regime clustering
# scikit-fuzzy>=0.4.2
pyyaml>=6.0
py7zr>=0.20.0
# yfinance>=0.2.0
